Refactor remove duplication
