+++ validators/index.js
@@ -1,0 +1,7 @@
+const validateString = require('./string')
+const validateNumber = require('./number')
+
+module.exports = {
+  validateString,
+  validateNumber
+}
